Remove InputLoopService implementation and refactor logic in Main
